pyproject.toml
[project]
name = "test-task-forwork"
version = "0.1.0"
description = ""
authors = [
    {name = "v_dmitrov",email = "v.dmitrov@smartmerch.it"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.14,<0.116.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
